git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@82583 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/domain/query/ibatis/CannedQueryDAOImpl.java
@@ -69,7 +69,7 @@ public class CannedQueryDAOImpl extends AbstractCannedQueryDAOImpl
         String query = makeQueryName(sqlNamespace, queryName);
         try
         {
-            Long result = (Long) template.selectOne(query, parameterObj);
+            Long result = template.selectOne(query, parameterObj);
             if (result == null)
             {
                 result = 0L;
